requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
qdrant-client==1.9.1
python-dotenv==1.0.1
pypdf==4.2.0
mlflow==2.10.2
ragas==0.1.7
datasets==2.17.0
pandas==2.0.3
numpy==1.24.4
tiktoken==0.7.0
openai==1.30.1
tqdm==4.66.4
scikit-learn==1.3.2
matplotlib==3.7.5
jupyter==1.0.0
ipykernel==6.29.4
httpx==0.27.2
sentence-transformers==2.2.2